Effects

+++ Assets/Scripts/Scriptables/Data/DataPoints/EnemyData.cs
@@ -6,8 +6,8 @@ namespace ScriptableArchitecture.Data
     [System.Serializable]
     public class EnemyData : IDataPoint
     {
-        public int Heath;
-        public int Damage;
+        public float Heath;
+        public float Damage;
         public float Speed;
         public int Coins;
         public Sprite Sprite;
